% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factonum.R
\name{factonum}
\alias{factonum}
\title{Fast converting from factors to numeric/integers}
\usage{
factonum(DT, cols, class = "numeric", removecomma = FALSE)
}
\arguments{
\item{DT}{The data.table to operate on}

\item{cols}{Quoted column names}

\item{class}{Which class to convert to: integer or numeric?}

\item{removecomma}{Should remove commas?}
}
\value{
The modified data.table with the refactored columns
}
\description{
Fast converting from factors to numeric/integers
}
\examples{
set.seed(123)
DT <- data.table(x = factor(sample(10)),
                 y = factor(sample(10) + round(runif(10), 3)),
                 z = factor(paste(sample(10), round(runif(10, 100, 1000), 3), sep = ",")),
                 l = factor(paste(sample(10), round(runif(10, 100, 1000), 3), sep = ",")))
DT
str(DT)
factonum(DT, "x", "integer")
factonum(DT, "y")
factonum(DT, c("z", "l"), removecomma = TRUE)
str(DT)
DT
}

