% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frollmean.R
\name{frollmean}
\alias{frollmean}
\title{Fast rolling mean}
\usage{
frollmean(DT, col, N, Name = "Mean", by, ...)
}
\arguments{
\item{DT}{The data.table to operate on}

\item{col}{A quoted column name}

\item{N}{An integer number that we want to shift by}

\item{Name}{(optional) the name of the new column}

\item{by}{(optional) A quoted by parameter}

\item{...}{Some additional parameters that you can pass to the \code{\link{shift}} function}
}
\value{
The modified data.table with the new shifted columns
}
\description{
Fast rolling mean
}
\examples{
set.seed(123)
DT <- data.table(x = sample(10), y = sample(1:2, 10, replace = TRUE), key = "y")
frollmean(DT, "x", 3, by = "y", type = "lead")
}

