% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frollprod.R
\name{frollprod}
\alias{frollprod}
\title{Fast rolling product}
\usage{
frollprod(DT, col, N, Name = "Prod", type = "lead", by)
}
\arguments{
\item{DT}{The data.table to operate on}

\item{col}{A quoted column name}

\item{N}{An integer number that we want to shift by}

\item{Name}{(optional) the name of the new column}

\item{type}{A string specifying the shift direction: "lag"/"lead"}

\item{by}{(optional) A quoted by parameter}
}
\value{
The modified data.table with the new shifted columns
}
\description{
Fast rolling product
}
\examples{
set.seed(123)
DT <- data.table(x = sample(10), y = sample(1:2, 10, replace = TRUE), key = "y")
frollprod(DT, "x", 3, by = "y")
}

