% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reducelevels.R
\name{reducelevels}
\alias{reducelevels}
\title{Reducing the amount of levels in factor columns}
\usage{
reducelevels(DT, cols, freq = 0.1, newlvl = "OTHER")
}
\arguments{
\item{DT}{The data.table to operate on}

\item{cols}{Quoted column names}

\item{freq}{The quantile from lower eqaul to you want to replace levels}

\item{newlvl}{The name of the new level}
}
\value{
The modified data.table with the refactored columns
}
\description{
Reducing the amount of levels in factor columns
}
\examples{
set.seed(123)
DT <- data.table(A = factor(sample(letters, 1e3L, replace = TRUE, prob = (1:length(letters))/100)),
                 B = factor(sample(letters, 1e3L, replace = TRUE, prob = rev(1:length(letters))/100)),
                 C = factor(sample(letters, 1e3L, replace = TRUE, prob = sample(1:length(letters))/100)),
                 D = sample(letters, 1e3, replace = TRUE),
                 E = sample(1e3))

str(DT)
reducelevels(DT)
str(DT)
reducelevels(DT, c("A", "B"), 0.6)
str(DT)
}

