% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{Fast removing unessacery levels from factor columns}
\usage{
refactor(DT, cols)
}
\arguments{
\item{DT}{The data.table to operate on}

\item{cols}{(optional) Quoted column names}
}
\value{
The modified data.table with the refactored columns
}
\description{
Fast removing unessacery levels from factor columns
}
\examples{
set.seed(123)
DT <- data.table(x = sample(10),
                 y = sample(1:2, 10, replace = TRUE),
                 z = factor(sample(letters, 10)),
                 l = factor(sample(LETTERS, 10)))

SDT <- DT[sample(5)]
str(SDT)
refactor(SDT)
str(SDT)
SDT <- DT[sample(5)]
str(SDT)
refactor(SDT, "z")
str(SDT)
}

